Implementación del principio HATEOAS - Parte 3

+++ api-rest-cuentas/src/main/java/com/api/hateoas/service/CuentaService.java
@@ -2,6 +2,7 @@ package com.api.hateoas.service;
 
 import com.api.hateoas.model.Cuenta;
 import com.api.hateoas.repository.CuentaRepository;
+import exception.CuentaNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -27,7 +28,10 @@ public class CuentaService {
         return cuentaRepository.save(cuenta);
     }
     
-    public void delete(Integer id){
+    public void delete(Integer id) throws CuentaNotFoundException{
+        if(!cuentaRepository.existsById(id)){
+            throw new CuentaNotFoundException("Cuenta no encontrada ID: " + id);
+        }
         cuentaRepository.deleteById(id);
     }
